# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  userId: string;
  displayName: string;
  email: string;
  referralCode: string;
  socialScore: number;
  coins: number;
  kycStatus: string;
  invitedCount: number;
  joinedCount: number;
  ptsEarned: number;
}

export interface UpdateProfileBody {
  displayName?: string;
  email?: string;
}

export interface Task {
  id: string;
  userId: string;
  title: string;
  type: string;
  status: string;
  location: string;
  dueDate: string;
  coins: number;
  icon: string;
}

export interface UpdateTaskBody {
  status?: string;
}