# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  userId: string;
  displayName: string;
  email: string;
  referralCode: string;
  socialScore: number;
  coins: number;
  kycStatus: string;
  invitedCount: number;
  joinedCount: number;
  ptsEarned: number;
}