# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Task {
  id: string;
  userId: string;
  title: string;
  type: string;
  status: string;
  location: string;
  dueDate: string;
  coins: number;
  icon: string;
}